+++ src/sections/FiveM.js
@@ -3,6 +3,12 @@ import { Button, ButtonGroup, Container, Grid, Typography } from '@material-ui/c
 import YouTubeEmbed from "../components/YouTubeEmbed"
 import Status from "../components/Status"
 import SocialMedia from '../components/SocialMedia';
+
+const getDate = () => {
+    let d = new Date();
+    return d.getFullYear() + "-" + (d.getMonth() + 1) + "-" + d.getDate()
+}
+
 export default {
     meta: {
         title: "FiveM Bot",
@@ -26,15 +32,11 @@ export default {
             ["Early-October 2021", "Revenue peaks at an all-time high."],
             ["October 28, 2021", "FiveM Bot obtains partnership with Snowside Hosting. Use code 'Zua' for 25% off recurring."],
             ["January 2022", "FiveM Bot is apart of 14K servers and has served over 5M users."]
-        ]
+        ],
     }, components: { statistics, links }
 }
 
 function statistics() {
-    let getDate = () => {
-        let d = new Date(); 
-        return d.getFullYear() + "-" + (d.getMonth()+1) + "-" + d.getDate()
-    }
     return (
         <Grid container spacing={1}>
             <Grid item>
@@ -55,15 +57,15 @@ function statistics() {
                 <img src="https://top.gg/api/widget/servers/483787947070586880.svg" alt="Discord Bots" />
                 <img src="https://top.gg/api/widget/owner/483787947070586880.svg" alt="Discord Bots" />
             </Grid>
-            <Grid style={{width: "100%", "minHeight": 500}} item>
+            <Grid style={{ width: "100%", "minHeight": 500 }} item>
                 <iframe
                     id="trends-widget-2"
                     src={`https://trends.google.com/trends/embed/explore/TIMESERIES?req=%7B%22comparisonItem%22%3A%5B%7B%22keyword%22%3A%22fivem%20bot%22%2C%22geo%22%3A%22%22%2C%22time%22%3A%222013-01-01%20${getDate()}%22%7D%5D%2C%22category%22%3A0%2C%22property%22%3A%22%22%7D&tz=240&forceMobileMode=false&isPreviewMode=true&eq=date%3Dall%26q%3Dfivem%2520bot&hl=enUS`}
                     width="100%"
                     height="450px"
-                    frameborder="0"
+                    frameBorder="0"
                     scrolling="0"
-                    style={{borderRadius: 5}}
+                    style={{ borderRadius: 5 }}
                     title="fivembot_trends"
                 />
             </Grid>
@@ -74,7 +76,7 @@ function statistics() {
 function links() {
     return (
         <div>
-            <SocialMedia icon="arrow-up-right-from-square" name="Use the bot" url="https://top.gg/bot/483787947070586880" confirmation/>
+            <SocialMedia icon="arrow-up-right-from-square" name="Use the bot" url="https://top.gg/bot/483787947070586880" confirmation />
             <YouTubeEmbed id="qphtBvXSusE" />
             <br />
             <Grid alignItems="baseline" container spacing={1}>
